CIHP_PGN(parsing)/requirements.txt
absl-py==0.13.0
astor==0.8.1
cycler==0.10.0
dataclasses==0.8
gast==0.5.2
grpcio==1.39.0
imageio==2.9.0
importlib-metadata==4.6.3
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.3.56
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.2
scipy==1.5.4
six==1.16.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
